refactor(AH02): fix indent in scss, rename func.ts in common.functions.ts
