+++ src/app/utils/functions/common.functions.ts
