+++ src/styles.scss
@@ -16,7 +16,6 @@ body {
   font-family: 'Roboto', sans-serif;
   font-size: 1rem;
   line-height: 1.6;
-  background-color: #fff;
   color: #333;
   background-color: #8ecae6;
 }
@@ -42,17 +41,17 @@ img {
 .container {
   max-width: 1100px;
   margin: auto;
-  overflow: hidden;
   padding: 0 2rem;
+  overflow: hidden;
 }
 
 .navbar {
   display: flex;
   align-items: center;
   justify-content: space-between;
+  margin-bottom: 1rem;
   padding: 0.7rem 2rem;
   width: 100%;
-  margin-bottom: 1rem;
   background-color: var(--primary-color);
 }
 
@@ -75,10 +74,10 @@ li {
   position: relative;
   width: 70px;
   height: 70px;
+  margin: 0 10px 0;
   border-radius: 50%;
   overflow: hidden;
   cursor: pointer;
-  margin: 0 10px 0;
 }
 
 img {
@@ -99,12 +98,13 @@ a {
 
 .form-control {
   display: block;
-  margin-bottom: .8rem;
+  margin-bottom: 0.8rem;
   color: #333;
 
   &.invalid {
     color: var(--danger-color);
     font-weight: bold;
+
     input {
       border: 4px solid var(--danger-color);
       border-color: var(--danger-color);
@@ -127,11 +127,13 @@ a {
   cursor: pointer;
   transition: opacity .2s ease-in;
   outline: none;
+
     &[disabled] {
       color: #111;
       background-color: var(--light-color);
       cursor: not-allowed;
     }
+
     &:hover {
       &:not([disabled]) {
         opacity: 0.6;
